homework/week_4/pyproject.toml
[project]
name = "week-4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dbt-bigquery>=1.11.0",
    "google-cloud-bigquery>=3.40.1",
    "google-cloud-storage>=3.1.1",
    "pandas>=2.3.3",
    "pyarrow>=23.0.0",
]
